awful.spawn: Use LGI's Async facilities
Normally, writing asynchronous code with Gio results in an awful mess of
spaghetti code, because every operation needs a callback that is called when
this operation is done. To improve upon this, LGI provides some magic with
coroutines that hide the asyncronous nature of things and you end up with
"normal looking" code.

Use this Gio.Async facility in awful.spawn to make the code easier to follow.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>